refactor: 修改maxGroupMemberCount属性类型[Integer->Long]

+++ src/main/java/com/r/chat/entity/dto/SysSettingDTO.java
@@ -22,7 +22,7 @@ public class SysSettingDTO implements Serializable, AvatarOwner {
     /**
      * 单个群聊最多群成员
      */
-    private Integer maxGroupMemberCount;
+    private Long maxGroupMemberCount;
 
     /**
      * 图片最大大小
